Changed/Fixed: overflowing long titles will now be truncated on list cards (tablet & desktop screens) and will scroll on hover to see them in full (like long list of tags)
